refactor: party entity 를 domain/entity 로 이동
