+++ src/main/kotlin/com/team2/server/chat/entity/ChatMessage.kt
@@ -1,8 +1,8 @@
 package com.team2.server.chat.entity
 
 import com.team2.server.common.persistence.BaseEntity
-import com.team2.server.party.entity.Party
-import com.team2.server.party.entity.RealtimeParticipantProfile
+import com.team2.server.party.domain.entity.Party
+import com.team2.server.party.domain.entity.RealtimeParticipantProfile
 import jakarta.persistence.Column
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
