+++ src/main/kotlin/com/team2/server/party/dto/CharacterImageUrlResolver.kt
@@ -2,7 +2,7 @@ package com.team2.server.party.dto
 
 import com.team2.server.common.image.entity.ImageTargetType
 import com.team2.server.common.image.persistence.ImageRepository
-import com.team2.server.party.entity.Character
+import com.team2.server.party.domain.entity.Character
 import org.springframework.stereotype.Component
 
 @Component
