+++ src/main/kotlin/com/team2/server/party/repository/CharacterRepository.kt
@@ -1,6 +1,6 @@
 package com.team2.server.party.repository
 
-import com.team2.server.party.entity.Character
+import com.team2.server.party.domain.entity.Character
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface CharacterRepository : JpaRepository<Character, Long> {
