+++ src/main/kotlin/com/team2/server/party/repository/ParticipantRepository.kt
@@ -1,7 +1,7 @@
 package com.team2.server.party.repository
 
-import com.team2.server.party.entity.Participant
-import com.team2.server.party.entity.Party
+import com.team2.server.party.domain.entity.Participant
+import com.team2.server.party.domain.entity.Party
 import com.team2.server.user.entity.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
