+++ src/main/kotlin/com/team2/server/party/repository/PartyInviteRepository.kt
@@ -1,6 +1,6 @@
 package com.team2.server.party.repository
 
-import com.team2.server.party.entity.PartyInvite
+import com.team2.server.party.domain.entity.PartyInvite
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
 import org.springframework.data.jpa.repository.Query
