+++ src/main/kotlin/com/team2/server/party/repository/RealtimeParticipantProfileRepository.kt
@@ -1,7 +1,7 @@
 package com.team2.server.party.repository
 
-import com.team2.server.party.entity.Participant
-import com.team2.server.party.entity.RealtimeParticipantProfile
+import com.team2.server.party.domain.entity.Participant
+import com.team2.server.party.domain.entity.RealtimeParticipantProfile
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Modifying
 import org.springframework.transaction.annotation.Transactional
