+++ src/main/kotlin/com/team2/server/rollingpaper/usecase/CreateRollingPaperUseCase.kt
@@ -3,8 +3,8 @@ package com.team2.server.rollingpaper.usecase
 import com.team2.server.common.exception.BusinessException
 import com.team2.server.common.exception.ErrorCode
 import com.team2.server.common.exception.isConstraintViolation
-import com.team2.server.party.entity.Participant
-import com.team2.server.party.entity.Party
+import com.team2.server.party.domain.entity.Participant
+import com.team2.server.party.domain.entity.Party
 import com.team2.server.party.service.ParticipantService
 import com.team2.server.party.service.PartyInviteService
 import com.team2.server.rollingpaper.dto.CreateRollingPaperRequest
