+++ src/main/kotlin/com/team2/server/rollingpaper/usecase/GetRollingPaperDetailUseCase.kt
@@ -2,7 +2,7 @@ package com.team2.server.rollingpaper.usecase
 
 import com.team2.server.common.exception.BusinessException
 import com.team2.server.common.exception.ErrorCode
-import com.team2.server.party.entity.Party
+import com.team2.server.party.domain.entity.Party
 import com.team2.server.party.repository.PartyRepository
 import com.team2.server.rollingpaper.dto.OwnerRollingPaperDetailResponse
 import com.team2.server.rollingpaper.entity.RollingPaper
