+++ src/test/kotlin/com/team2/server/chat/controller/ChatControllerTest.kt
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.team2.server.auth.config.JwtProperties
 import com.team2.server.auth.jwt.JwtTokenProvider
 import com.team2.server.chat.repository.ChatMessageRepository
-import com.team2.server.party.entity.Character
-import com.team2.server.party.entity.PartyInvite
-import com.team2.server.party.entity.RealtimeParty
+import com.team2.server.party.domain.entity.Character
+import com.team2.server.party.domain.entity.PartyInvite
+import com.team2.server.party.domain.entity.RealtimeParty
 import com.team2.server.party.repository.CharacterRepository
 import com.team2.server.party.repository.ParticipantRepository
 import com.team2.server.party.repository.PartyInviteRepository
@@ -95,7 +95,7 @@ class ChatControllerTest
             val character = characterRepository.save(Character(name = "bird"))
             val party =
                 partyRepository.save(
-                    com.team2.server.party.entity.PaperOnlyParty(
+                    com.team2.server.party.domain.entity.PaperOnlyParty(
                         ownerId = owner.id,
                         celebrantNickname = "홍길동",
                         startedAt = LocalDateTime.now().plusHours(1),
