+++ src/test/kotlin/com/team2/server/party/domain/entity/PartyDomainTest.kt
@@ -1,4 +1,4 @@
-package com.team2.server.party.entity
+package com.team2.server.party.domain.entity
 
 import org.junit.jupiter.api.Test
 import java.time.LocalDateTime
