+++ src/test/kotlin/com/team2/server/rollingpaper/entity/RollingPaperDomainTest.kt
@@ -1,8 +1,8 @@
 package com.team2.server.rollingpaper.entity
 
-import com.team2.server.party.entity.Participant
-import com.team2.server.party.entity.Party
-import com.team2.server.party.entity.RealtimeParty
+import com.team2.server.party.domain.entity.Participant
+import com.team2.server.party.domain.entity.Party
+import com.team2.server.party.domain.entity.RealtimeParty
 import org.junit.jupiter.api.Test
 import java.time.LocalDateTime
 import kotlin.test.assertEquals
